refactor(web): move run creation helper to api-test-helpers
- add helper for creating runs via internal API
- remove local helper function from route.test.ts
- follows same pattern as other test helpers
